+++ js/modules/gui/utils.mjs
@@ -1,4 +1,4 @@
-import { settings, gStyle, isBatchMode, isNodeJs, isFunc, isStr, source_dir, atob_func, btoa_func } from '../core.mjs';
+import { settings, gStyle, isBatchMode, isNodeJs, isObject, isFunc, isStr, source_dir, atob_func, btoa_func } from '../core.mjs';
 import { select as d3_select, pointer as d3_pointer, drag as d3_drag, color as d3_color } from '../d3.mjs';
 import { BasePainter } from '../base/BasePainter.mjs';
 import { resize } from '../base/ObjectPainter.mjs';
@@ -405,7 +405,7 @@ function readCookie(name) {
       if (c.indexOf(name) == 0) {
          let s = JSON.parse(atob_func(c.substring(name.length, c.length)));
 
-         return (s && typeof s == 'object') ? s : null;
+         return isObject(s) ? s : null;
       }
    }
    return null;