+++ js/changes.md
@@ -11,6 +11,7 @@
 8. Fix - interactive move/resize on touch devices
 9. Provide `settings.PreferSavedPoints` to exclude function evaluation when there are saved points
 10. Improve TWebCanvas support
+11. Correctly handle axis zooming on lego plots
 
 
 ## Changes in 7.2.1